efi.c,efi.h: Add mouse test via Simple Pointer Protocol
Added a menu option to test the mouse, which gets simple pointer
protocol (SPP) instances from available handles with LocateHandleBuffer()
and prints the SPP info and mouse state. Also am drawing a sample
"cursor" for a square on the screen, which is bounded to the GOP
mode horizontal/vertical resolution.

There are probably ways to have smoother and more accurate mouse
movement, which I should research, but for basic mouse support
it seems to work fine on my laptop.

Unfortunately qemu/ovmf does not include a working driver for a
pointer device to use the SPP in emulation. I would have to build
a custom ovmf image, and online resources are lacking on how
to do that specifically to include a DXE and/or mouse driver.