Add file protocols, read through ESP and print files.
Added Loaded Image, Simple File System, and File protocols,
as well as Allocate Pool, to be able to read the EFI System
Partition (ESP) files and directories from the running EFI
application. The ESP is normally a FAT filesystem.

Added a menu option to read ESP files, with up/down arrows to
choose a directory entry and enter key to enter to enter a
nested/parent directory or print a file.