Read file/data from Basic Data Partition.
Added menu option that will be used to load a binary file for a
"kernel" in the future. Currently it loads a test text file from
the data partition on the running disk image.

Added various helper functions for better abstractions and a couple
libc functions to help with this.