Add program loading from 'Load Kernel' menu option.
Change Load Kernel option to load an example "kernel",
which takes in the GOP framebuffer info and draws to screen.
Allow creating program as an ELF64, PE32+, or 64 bit flat binary
file. All formats are (should) be generated as Position Independant
Executables (PIE). This makes loading fairly easy, as well as the
program being simple itself, at the moment.

Choose in the makefile by uncommenting which kernel format to build:
elf, pe, or bin.