Use linker script for flat binary kernel.
Not needed for ELF/PE, only plain binary to explicity set the entry
point at the start of executable code (.text section). Using a
separate section .kernel and using KEEP in a linker script to ensure
it is before the .text section proper, at the start of executable code
and the flat binary object itself.

For mingw making a PE object into the flat binary, set the image base
to 0 when linked for a simple fix for relocation errors.

All binary/ELF/PE kernels should be good to go now, I hope...