Removed some mapping and calling runtime SetVirtualAddressMap.
For only loading the kernel, did not need to identity map
each new entry in each page table in map_page(). Also did not
need to set VirtualStart = Physical start in EFI memory map entries
and did not need to call Runtime Services SetVirtualAddressMap().

Those may be useful or good to do, if you want to be able to call
runtime services from the loaded OS, but aren't needed for simply
loading the kernel in this example.