Use AllocatePages() for disk lba buffer for page aligned address.
Change read_disk_lbas_to_buffer() to return a pointer to a page
aligned address allocated with AllocatePages(), not AllocatePool().
This fixes loading a flat binary kernel with an entry point at the
start of the allocated buffer, and allows mapping the loaded kernel
buffer to any other page aligned addresses.

Before, allocate pool returned a byte aligned buffer, not page aligned,
and the flat binary entry point was byte aligned from this.
Mapping pages to a page aligned address wipes out the lowest 12 bits
used for flags, and those 12 bits held part of the byte aligned address,
so that the kernel entry point was not valid and caused a page fault
when jumping to the kernel.

This fixes that by always having a page aligned address for the loaded
kernel buffer and entry point, flat binary or not.