Move helper vars/functions to efi_lib.h
Move some global variables and general helper functions to
efi_lib.h, as a "library" header file. Helps to clean up
the main efi.c a little bit.

Add basic ELF/PE file printing functions to efi_lib.h.

Add -fno-stack-protector to CFLAGS in makefile, as clang was trying
to link in a stack check routine but with -ffreestanding the host
libc is not going to be found or used.