Add "install" option to load kernel instead of main menu.
Added helper function to return EFI System Partition root '/'
directory.

Added menu option & function to create new empty file
/EFI/BOOT/INSTALL.DAT, to signal the disk image or boot loader
has been "installed". This file is checked on startup in efi_main()
and if it exists then load_kernel() is called automatically,
instead of loading the main menu.

It is also called when writing the disk image to another disk, so
that on the first boot from that new disk, it can autoload the
kernel as well.

- Thanks to ThatOSDev for install file and efi_lib.h file suggestions,
    and conversation.