Use default AARCH64 OVMF EFI/bios for smaller repo size.
Add makefile macros for "detecting" the host os and setting
the arch specific ovmf bios to use.
Assuming Linux or Windows (if windows works at this point),
and fsutil on windows as a "dd" equivalent.

Not needed if using qemu -bios flag, but qemu aarch64 seemed to
need 64MiB files for pflash devices instead.